Don't delete non-existing battle

+++ index_actions.php
@@ -6,7 +6,7 @@ $battle_id = get('id');
 
 $battle = $battle_id ? Battle::from_id($battle_id) : NULL;
 
-if(get("action") == "Radera") {
+if($battle && get("action") == "Radera") {
 	$battle->delete();
 	redirect("index.php");
 }
